+++ samples/apps/cap/py/autogencap/DirectorySvc.py
@@ -8,7 +8,7 @@ from autogencap.Actor import Actor
 from autogencap.ActorConnector import ActorConnector, ActorSender
 from autogencap.Broker import Broker
 from autogencap.Config import router_url, xpub_url, xsub_url
-from autogencap.Constants_ import Directory_Svc_Topic
+from autogencap.constants import Directory_Svc_Topic
 from autogencap.DebugLog import Debug, Error, Info
 from autogencap.proto.CAP_pb2 import (
     ActorInfo,
