+++ samples/apps/cap/py/autogencap/runtime_factory.py
@@ -1,5 +1,5 @@
 from autogencap.actor_runtime import IRuntime
-from autogencap.Constants_ import ZMQ_Runtime
+from autogencap.constants import ZMQ_Runtime
 from autogencap.DebugLog import Error
 from autogencap.zmq_runtime import ZMQRuntime
